app/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
requests>=2.0
django-environ>=0.4
djangorestframework>=3.12
django-import-export>=2.5
django-cors-headers>=3.7
